Create auth.php filter and use them for several pages

+++ pages/home.php
@@ -1,5 +1,10 @@
 <?php 
-  
+  $pseudo = get_session_user();
+  $auth = false;
+  if (isset($pseudo) && !empty($pseudo) ) {
+      $auth = true;
+  }
+
   if (isset($_GET['search'])) {
     $results = get_books_all($_GET['search']);
     $books = $results['rows'];
@@ -29,7 +34,11 @@
               <th>Titre</th>
               <th>Ecrivain</th>
               <th>Publié le</th>
-              <th>Actions</th>
+              <?php 
+                if($auth) { ?>
+                  <th>Actions</th>
+              <?php  }
+              ?>
           </tr>
       </thead>
       <tbody>
@@ -42,12 +51,17 @@
                       <td><?= $book->title ?></td>
                       <td><?= strtolower($book->firstname.' '.$book->lastname); ?></td>
                       <td><?= $book->publish_at ?></td>
-                      <td>
-                        <div class="areaActions">
-                          <a class='btn-modify' href="?page=update-book&id=<?= $id ?>"><i class="fas fa-edit"></i></a>
-                          <a  class='btn-delete' data-id="<?= $id ?>" href="#" onclick="deleteElement(this)"><i class="fas fa-trash-alt"></i></a>
-                        </div>
-                      </td>
+                      <?php 
+                        if($auth) { ?>
+                          <td>
+                            <div class="areaActions">
+                              <a class='btn-modify' href="?page=update-book&id=<?= $id ?>"><i class="fas fa-edit"></i></a>
+                              <a  class='btn-delete' data-id="<?= $id ?>" href="#" onclick="deleteElement(this)"><i class="fas fa-trash-alt"></i></a>
+                            </div>
+                          </td>
+                      <?php  } 
+                      ?>
+                      
                   </tr>
           <?php }
           ?>
